Refactoring all the new nodes and agents under sotopia/experimental
