+++ sotopia/experimental/agents/agents.py
@@ -1,17 +1,63 @@
 import asyncio
 import sys
 
+from enum import Enum
+from pydantic import Field
+
 if sys.version_info < (3, 11):
     from typing_extensions import Self
 else:
     from typing import Self
 
-from typing import Any, AsyncIterator, TypeVar
-from aact import Node
-from aact.messages import DataModel, Message
+from typing import Any, AsyncIterator, TypeVar, Optional
+
+from aact import Message, Node
+from aact.messages import DataModel, DataModel, Message
+from aact.messages.registry import DataModelFactory
 
 T_agent_observation = TypeVar("T_agent_observation", bound=DataModel)
 T_agent_action = TypeVar("T_agent_action", bound=DataModel)
+
+
+class ActionType(Enum):
+    NONE = "none"
+    SPEAK = "speak"
+    NON_VERBAL_COMMUNICATION = "non-verbal communication"
+    ACTION = "action"
+    LEAVE = "leave"
+    THOUGHT = "thought"
+    BROWSE = "browse"
+    BROWSE_ACTION = "browse_action"
+    READ = "read"
+    WRITE = "write"
+    RUN = "run"
+
+@DataModelFactory.register("agent_action")
+class AgentAction(DataModel):
+    agent_name: str = Field(description="the name of the agent")
+    action_type: ActionType = Field(
+        description="whether to speak at this turn or choose to not do anything"
+    )
+    argument: str = Field(
+        description="the utterance if choose to speak, the expression or gesture if choose non-verbal communication, or the physical action if choose action"
+    )
+    path: Optional[str] = Field(description="path of file")
+
+    def to_natural_language(self) -> str:
+        action_descriptions = {
+            ActionType.NONE: "did nothing",
+            ActionType.SPEAK: f'said: "{self.argument}"',
+            ActionType.THOUGHT: f'thought: "{self.argument}"',
+            ActionType.BROWSE: f'browsed: "{self.argument}"',
+            ActionType.RUN: f'ran: "{self.argument}"',
+            ActionType.READ: f'read: "{self.argument}"',
+            ActionType.WRITE: f'wrote: "{self.argument}"',
+            ActionType.NON_VERBAL_COMMUNICATION: f"[{self.action_type.value}] {self.argument}",
+            ActionType.ACTION: f"[{self.action_type.value}] {self.argument}",
+            ActionType.LEAVE: "left the conversation",
+        }
+
+        return action_descriptions.get(self.action_type, "performed an unknown action")
 
 
 class BaseAgent(Node[T_agent_observation, T_agent_action]):
